test: refactor and expand picker test suite
